requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
PyJWT>=2.8
jwcrypto>=1.4
cryptography>=41.0
python-dotenv>=1.0
requests>=2.31
psycopg2-binary>=2.9
bcrypt~=5.0.0
pip~=21.3.1
wheel~=0.37.1
setuptools~=60.2.0
asgiref~=3.7.2
sqlparse~=0.4.4
psycopg~=3.1.12
certifi~=2023.7.22
urllib3~=2.0.7
idna~=3.4
pycparser~=2.21
Markdown~=3.4.3
pytz~=2023.3